requirement.txt
--index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn
transformers
pillow
python-multipart
